Split library into multi-module architecture (core, rest, jdbc)
- Add build-logic with precompiled Gradle script plugins
- Move core models (Pageable, Page, PaginationError) to core module
- Move Ktor integration (PageableRequest) to rest module
- Move Exposed extensions (PageableQuery, MapModel) to jdbc module
