+++ core/src/main/kotlin/io/perracodex/exposed/pagination/Page.kt
