+++ core/src/main/kotlin/io/perracodex/exposed/pagination/PaginationError.kt
