+++ gradle/libs.versions.toml
@@ -5,6 +5,8 @@
 detekt = "1.23.7"
 dokka = "2.0.0"
 exposed = "1.0.0-rc-4"
+h2 = "2.3.232"
+kotest = "6.0.0.M1"
 kotlin = "2.2.21"
 kotlinx-serialization = "1.9.0"
 ktor = "3.3.2"
@@ -64,3 +66,19 @@ kotlinx-serialization = { module = "org.jetbrains.kotlinx:kotlinx-serialization-
 # https://github.com/ktorio/ktor
 # https://ktor.io/docs/releases.html#release-details
 ktor-server-core = { module = "io.ktor:ktor-server-core", version.ref = "ktor" }
+
+# Testing libraries.
+# https://github.com/kotest/kotest/releases
+# https://github.com/kotest/kotest
+kotest-runner-junit5 = { module = "io.kotest:kotest-runner-junit5", version.ref = "kotest" }
+kotest-assertions-core = { module = "io.kotest:kotest-assertions-core", version.ref = "kotest" }
+
+# H2 Database.
+# https://github.com/h2database/h2database/releases
+h2 = { module = "com.h2database:h2", version.ref = "h2" }
+
+# Ktor Test Host.
+ktor-server-test-host = { module = "io.ktor:ktor-server-test-host", version.ref = "ktor" }
+ktor-client-content-negotiation = { module = "io.ktor:ktor-client-content-negotiation", version.ref = "ktor" }
+ktor-server-content-negotiation = { module = "io.ktor:ktor-server-content-negotiation", version.ref = "ktor" }
+ktor-serialization-kotlinx-json = { module = "io.ktor:ktor-serialization-kotlinx-json", version.ref = "ktor" }
